requirements.txt
aiohttp==3.12.4
beautifulsoup4==4.13.4
blessed==1.21.0
lxml==5.4.0
pandas==2.3.0
Pillow==11.2.1
psycopg2==2.9.10
python-dotenv==1.1.0
Requests==2.32.3
